requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.1.0
email-validator==2.1.0
Flask-Mail==0.9.1

